reorganized, created sql files, updated README
